Create Fibonacci_Serie
Fibonacci designed his series to model the reproduction speed of rabbits. His model is based on adding the initial
two numbers of the series to generate the next number. Example: given numbers 0 and 1 we add those together to come
to the next number in the series 0 + 1 = 1, then we add that new number to the previous number in the series
1 + 1 = 2, and again 1 + 2 = 3, and again 3 + 2 = 5, and again 5 + 3 = 8, and so on...