Create Factorial.py
Factorial is a mathematical task represented by the exclamation sign (" ! "). It represents the product of an
integer and all the integers below it; e.g., factorial four ( 4! ) is equal to 24 and factorial five ( 5! ) is equal to 120. Per its definition, factorial
only involves POSITIVE integers (no negative or float point values)